--- src/main/java/com/SurveyMonkey/controller/ViewController.java
@@ -1,14 +1,0 @@
-package com.SurveyMonkey.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class ViewController {
-
-	@GetMapping("/createSurveyQuestion")
-	public String home() {
-		return "createSurveyQuestion";
-	}
-
-}
