+++ src/main/java/com/SurveyMonkey/controller/rest/MultipleChoiceQuestionController.java
@@ -1,9 +1,11 @@
-package com.SurveyMonkey.controller;
+package com.SurveyMonkey.controller.rest;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import com.SurveyMonkey.Service.ServiceResponse;
+import com.SurveyMonkey.repositories.CreateSurveyRepository;
+import com.SurveyMonkey.repositories.MultipleChoiceRepository;
+import com.SurveyMonkey.controller.response.ServiceResponse;
 import com.SurveyMonkey.model.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
