+++ src/main/java/com/SurveyMonkey/model/DropDownQuestion.java
@@ -1,11 +1,6 @@
 package com.SurveyMonkey.model;
 
-import com.SurveyMonkey.model.questions.types.QType;
-
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 
 @Entity
 public class DropDownQuestion extends QType {
