+++ src/main/java/com/SurveyMonkey/model/MultipleChoiceQuestionModel.java
@@ -1,7 +1,5 @@
 package com.SurveyMonkey.model;
 
-import com.SurveyMonkey.model.questions.types.QType;
-
 import javax.persistence.Entity;
 
 @Entity
