--- src/main/java/com/SurveyMonkey/model/answers/types/ARadio.java
@@ -1,13 +1,0 @@
-package com.SurveyMonkey.model.answers.types;
-
-public class ARadio extends AType {
-    public String answer;
-
-    public String getAnswer() {
-        return answer;
-    }
-
-    public void setAnswer(String answer) {
-        this.answer = answer;
-    }
-}
