--- src/main/java/com/SurveyMonkey/model/answers/types/AType.java
@@ -1,4 +1,0 @@
-package com.SurveyMonkey.model.answers.types;
-
-public abstract class AType {
-}
