+++ src/main/java/com/SurveyMonkey/repositories/CreateSurveyRepository.java
@@ -1,4 +1,4 @@
-package com.SurveyMonkey.model;
+package com.SurveyMonkey.repositories;
 
 import com.SurveyMonkey.model.CreateSurvey;
 import org.springframework.data.repository.CrudRepository;
