+++ src/main/java/com/SurveyMonkey/repositories/OpenEndedQuestionRepository.java
@@ -1,5 +1,6 @@
-package com.SurveyMonkey.model;
+package com.SurveyMonkey.repositories;
 
+import com.SurveyMonkey.model.OpenEndedQuestion;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
