+++ src/main/resources/static/css/confirmationPage.css
@@ -1,0 +1,19 @@
+body{
+    text-align: center;
+}
+.button {
+    background-color: #4CAF50;
+    border: none;
+    color: white;
+    padding: 15px 32px;
+    text-align: center;
+    text-decoration: none;
+    display: inline-block;
+    font-size: 16px;
+    margin: 4px 2px;
+    cursor: pointer;
+}
+img{
+    width: 10%;
+    height: 10%;
+}
