+++ src/main/resources/static/css/takerSurveyPage.css
@@ -1,0 +1,27 @@
+div.element {
+    margin-left:auto;
+    margin-right:auto;
+    border-collapse: separate;
+    border-spacing: 10px 10px;
+    border: 1px solid black;
+    margin-bottom: 25px;
+    width: 500px;
+    border-radius: 15px;
+    justify-content: center;
+
+}
+text.questionText {
+    padding: 10px;
+    text-align: left;
+
+}
+h3 {
+    color: #1b8c39;
+}
+h4 {
+    text-align: center;
+    color: #7e807e
+}
+h1, h2{
+    text-align: center;
+}
