+++ src/main/resources/static/scripts/getRequestForOpenEndedQuestion.js
