+++ src/main/resources/static/scripts/getrequest.js
