+++ src/main/resources/static/scripts/postRequestForOpenEndedQuestion.js
