+++ src/main/resources/templates/General.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>General Page</title>
-    <link rel="stylesheet" type="text/css" href="styles/styles.css" />
+    <link rel="stylesheet" type="text/css" href="../static/css/styles.css" />
 </head>
 <body>
