+++ src/main/resources/templates/confirmationPage.html
@@ -4,29 +4,9 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Document</title>
+  <title>Confirm</title>
+  <link href="../static/css/confirmationPage.css" rel="stylesheet" type="text/css"/>
 </head>
-<style>
-  body{
-    text-align: center;
-  }
-  .button {
-    background-color: #4CAF50;
-    border: none;
-    color: white;
-    padding: 15px 32px;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    font-size: 16px;
-    margin: 4px 2px;
-    cursor: pointer;
-  }
-  img{
-    width: 10%;
-    height: 10%;
-  }
-  </style>
 <body>
   <h1>Survey Title</h1>
   <br>
