+++ src/main/resources/templates/createSurvey.html
@@ -3,7 +3,7 @@
 <head>
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
     <title>Surveyor: Create Survey Page</title>
-    <link rel="stylesheet" type="text/css" href="styles/styles.css" />
+    <link rel="stylesheet" type="text/css" href="../static/css/styles.css" />
     <script type="text/javascript" src="/scripts/createSurvey.js"></script>
 </head>
 <body>
