+++ src/main/resources/templates/surveyResults.html
@@ -3,39 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Survey Results</title>
-    <style>
-        table {
-            border-collapse: separate;
-            border-spacing: 0 1px;
-            border-top: 1px solid black;
-            border-right: 1px solid black;
-            border-left: 1px solid black;
-        }
-        td {
-            text-align: center;
-            width: 600px;
-            padding: 1px;
-            border-bottom: 1pt solid black;
-        }
-        tr {
-            padding: 2px;
-            border-color: #96D4D4;
-            border-bottom: 1pt solid black
-        }
-        h3 {
-            color: #000000;
-        }
-        h4 {
-            text-align: center;
-            color: #7e807e
-        }
-        h1, h2{
-            text-align: center;
-        }
-        .chartBox {
-            width: 600px;
-        }
-    </style>
+    <link href="../static/css/surveyResults.css" rel="stylesheet" type="text/css"/>
 </head>
 <body>
 <div style="justify-content: center;text-align:center;">
