+++ src/main/resources/templates/surveyorEdit.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Edit Survey</title>
-    <link href="styles/styles.css" rel="stylesheet" type="text/css"/>
+    <link href="../static/css/styles.css" rel="stylesheet" type="text/css"/>
 </head>
 
 <body>
