+++ src/test/java/com/SurveyMonkey/SurveyMonkeyApplicationHttpRequestTest.java
@@ -11,7 +11,7 @@ import org.springframework.boot.web.server.LocalServerPort;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-public class HttpRequestTest {
+public class SurveyMonkeyApplicationHttpRequestTest {
 
     @LocalServerPort
     private int port;
