+++ src/test/java/com/SurveyMonkey/controller/HomePageControllerTests.java
@@ -1,6 +1,8 @@
-package com.SurveyMonkey;
+package com.SurveyMonkey.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
+
+import com.SurveyMonkey.controller.HomePageController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
