+++ src/test/java/com/SurveyMonkey/controller/SurveyControllerTests.java
@@ -1,4 +1,4 @@
-package com.SurveyMonkey;
+package com.SurveyMonkey.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
